Add ability to query for results
