+++ packages/backend/src/simulation/simulation.resolver.ts
@@ -1,8 +1,10 @@
-import { Args, Resolver, Query, Mutation } from '@nestjs/graphql';
+import { Args, Resolver, Query, Mutation, ID } from '@nestjs/graphql';
 import { SimulationRequestInput } from './dtos/request.dto';
+import { SimulationRequest } from './models/request.model';
+import { SimulationPipe } from './pipes/simulation.pipe';
 import { SimulationService } from './simulation.service';
 
-@Resolver()
+@Resolver(() => SimulationRequest)
 export class SimulationResolver {
   constructor(private readonly simulationService: SimulationService) {}
 
@@ -15,5 +17,10 @@ export class SimulationResolver {
   @Query(() => String)
   async example(): Promise<string> {
     return 'hello';
+  }
+
+  @Query(() => SimulationRequest)
+  async getSimulationRequest(@Args('request', { type: () => ID }, SimulationPipe) request: SimulationRequest): Promise<SimulationRequest> {
+    return request;
   }
 }
